* Fixed the hashing so it properly starts and stops. When DownloadList::pause is called it clears the "hashing" variable.
* Added convenience functions for retriving torrent::Object key's of
specific types.

* Iterate to the first visible transfer in the chunks seen view. Patch
by Josef Drexler.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/rtorrent@732 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/core/download.h
@@ -55,10 +55,10 @@ public:
   typedef download_type::ConnectionType connection_type;
   typedef utils::VariableMap            variable_map_type;
 
-  static const int64_t variable_hashing_stopped = 0;
-  static const int64_t variable_hashing_initial = 1;
-  static const int64_t variable_hashing_last    = 2;
-  static const int64_t variable_hashing_rehash  = 3;
+  static const int variable_hashing_stopped = 0;
+  static const int variable_hashing_initial = 1;
+  static const int variable_hashing_last    = 2;
+  static const int variable_hashing_rehash  = 3;
 
   Download(download_type d);
   ~Download();
